feat(snapshot): review comment fix, move logical_volume outside lvs
