+++ io-engine/src/grpc/v0/mayastor_grpc.rs
@@ -13,6 +13,7 @@ use crate::{
     bdev_api::BdevError,
     core::{
         lock::{ProtectedSubsystems, ResourceLockManager},
+        logical_volume::LogicalVolume,
         BlockDeviceIoStats,
         CoreError,
         MayastorFeatures,
@@ -39,13 +40,7 @@ use crate::{
         Serializer,
     },
     host::{blk_device, resource},
-    lvs::{
-        lvs_lvol::{LogicalVolume, SpdkLvol},
-        Error as LvsError,
-        Lvol,
-        LvolSpaceUsage,
-        Lvs,
-    },
+    lvs::{lvs_lvol::LvsLvol, Error as LvsError, Lvol, LvolSpaceUsage, Lvs},
     pool_backend::PoolArgs,
     rebuild::{RebuildRecord, RebuildState, RebuildStats},
     subsys::PoolConfig,
