+++ io-engine/src/lvs/mod.rs
@@ -1,14 +1,7 @@
 pub use lvs_bdev::LvsBdev;
 pub use lvs_error::Error;
 pub use lvs_iter::{LvsBdevIter, LvsIter};
-pub use lvs_lvol::{
-    LogicalVolume,
-    Lvol,
-    LvolSpaceUsage,
-    PropName,
-    PropValue,
-    SpdkLvol,
-};
+pub use lvs_lvol::{Lvol, LvolSpaceUsage, LvsLvol, PropName, PropValue};
 pub use lvs_store::Lvs;
 
 mod lvs_bdev;
