+++ io-engine/src/subsys/nvmf/admin_cmd.rs
@@ -9,11 +9,8 @@ use std::{
 
 use crate::{
     bdev::nexus,
-    core::{Bdev, Reactors},
-    lvs::{
-        lvs_lvol::{LogicalVolume, SpdkLvol},
-        Lvol,
-    },
+    core::{logical_volume::LogicalVolume, Bdev, Reactors},
+    lvs::{lvs_lvol::LvsLvol, Lvol},
 };
 
 use spdk_rs::{
